Encode "+" as "%2B" in username and password
QUrlQuery performs URL encoding, except for encoding of spaces and +
symbols as would be the case for an HTML form. See:

https://doc.qt.io/qt-5/qurlquery.html#handling-of-spaces-and-plus

Hacker News seems to expect the "+" symbols to be encoded.

As a result, encoding of plus symbols must be performed manually in
order for usernames and passwords that include the symbol to work
(spaces seem to pass through okay if left as spaces).